web_server/requirements.txt
flask==3.0.2
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
google-generativeai==0.7.1